# Speed up AST rewriting