--- a/go/tools/asthelpergen/asthelpergen.go
+++ b/go/tools/asthelpergen/asthelpergen.go
@@ -23,6 +23,7 @@ import (
 	"io/ioutil"
 	"log"
 	"path"
+	"sort"
 	"strings"
 
 	"github.com/dave/jennifer/jen"
@@ -43,37 +44,59 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.`
 
-type generatorSPI interface {
-	addType(t types.Type)
-	addFunc(name string, t methodType, code jen.Code)
-	scope() *types.Scope
-	findImplementations(iff *types.Interface, impl func(types.Type) error) error
-	iface() *types.Interface
+type (
+	generatorSPI interface {
+		addType(t types.Type)
+		addFunc(name string, t methodType, code jen.Code)
+		scope() *types.Scope
+		findImplementations(iff *types.Interface, impl func(types.Type) error) error
+		iface() *types.Interface
+	}
+	generator2 interface {
+		interfaceMethod(t types.Type, iface *types.Interface, spi generatorSPI) error
+		structMethod(t types.Type, strct *types.Struct, spi generatorSPI) error
+		ptrToStructMethod(t types.Type, strct *types.Struct, spi generatorSPI) error
+		ptrToBasicMethod(t types.Type, basic *types.Basic, spi generatorSPI) error
+		sliceMethod(t types.Type, slice *types.Slice, spi generatorSPI) error
+		basicMethod(t types.Type, basic *types.Basic, spi generatorSPI) error
+	}
+	// astHelperGen finds implementations of the given interface,
+	// and uses the supplied `generator`s to produce the output code
+	astHelperGen struct {
+		DebugTypes bool
+		mod        *packages.Module
+		sizes      types.Sizes
+		namedIface *types.Named
+		_iface     *types.Interface
+		gens       []generator2
+
+		functions methods
+		_scope    *types.Scope
+		todo      []types.Type
+	}
+
+	method struct {
+		name string
+		code jen.Code
+		typ  methodType
+	}
+
+	methods []method
+)
+
+func (m methods) Len() int {
+	return len(m)
 }
 
-type generator2 interface {
-	interfaceMethod(t types.Type, iface *types.Interface, spi generatorSPI) error
-	structMethod(t types.Type, strct *types.Struct, spi generatorSPI) error
-	ptrToStructMethod(t types.Type, strct *types.Struct, spi generatorSPI) error
-	ptrToBasicMethod(t types.Type, basic *types.Basic, spi generatorSPI) error
-	sliceMethod(t types.Type, slice *types.Slice, spi generatorSPI) error
-	basicMethod(t types.Type, basic *types.Basic, spi generatorSPI) error
+func (m methods) Less(i, j int) bool {
+	return m[i].name < m[j].name
 }
 
-// astHelperGen finds implementations of the given interface,
-// and uses the supplied `generator`s to produce the output code
-type astHelperGen struct {
-	DebugTypes bool
-	mod        *packages.Module
-	sizes      types.Sizes
-	namedIface *types.Named
-	_iface     *types.Interface
-	gens       []generator2
-
-	methods []jen.Code
-	_scope  *types.Scope
-	todo    []types.Type
+func (m methods) Swap(i, j int) {
+	m[i], m[j] = m[j], m[i]
 }
+
+var _ sort.Interface = (methods)(nil)
 
 func (gen *astHelperGen) iface() *types.Interface {
 	return gen._iface
@@ -264,18 +287,7 @@ const (
 )
 
 func (gen *astHelperGen) addFunc(name string, typ methodType, code jen.Code) {
-	var comment string
-	switch typ {
-	case clone:
-		comment = " creates a deep clone of the input."
-	case equals:
-		comment = " does deep equals between the two objects."
-	case visit:
-		comment = " will visit all parts of the AST"
-	case rewrite:
-		comment = " is part of the Rewrite implementation"
-	}
-	gen.methods = append(gen.methods, jen.Comment(name+comment), code)
+	gen.functions = append(gen.functions, method{name: name, code: code, typ: typ})
 }
 
 func (gen *astHelperGen) createFile(pkgName string) (string, *jen.File) {
@@ -334,8 +346,16 @@ func (gen *astHelperGen) createFile(pkgName string) (string, *jen.File) {
 		alreadyDone[typeName] = true
 	}
 
-	for _, method := range gen.methods {
-		out.Add(method)
+	sort.Sort(gen.functions)
+
+	for _, m := range gen.functions {
+		switch m.typ {
+		case clone:
+			out.Add(jen.Comment(fmt.Sprintf("%s creates a deep clone of the input.", m.name)))
+		case equals:
+			out.Add(jen.Comment(fmt.Sprintf("%s does deep equals between the two objects.", m.name)))
+		}
+		out.Add(m.code)
 	}
 
 	return "ast_helper.go", out