… .github/workflows/check_make_visitor.yml
… go/tools/asthelpergen/integration/integration_rewriter_test.go
… go/tools/asthelpergen/integration/rewriter.go
… go/tools/asthelpergen/integration/types.go
… go/tools/asthelpergen/rewrite_gen.go
… go/tools/asthelpergen/rewriter_gen.go
… go/tools/asthelpergen/visit_gen.go
… go/vt/sqlparser/ast_helper.go
… go/vt/sqlparser/rewriter.go
… go/vt/sqlparser/rewriter_api.go
… go/vt/sqlparser/rewriter_test.go
… go/vt/sqlparser/walker_test.go
… misc/git/hooks/asthelpers